refactor: ポリシーページをコンポーネント分離・整理(#55)
